+++ sora_stats_webhook.go
@@ -8,19 +8,10 @@ import (
 )
 
 // ログレベル、ログメッセージを変更する
-func (s *Server) collector(c echo.Context) error {
-	if c.Request().ProtoMajor != 2 {
-		zlog.Error().
-			Str("Proto", c.Request().Proto).
-			Int("ProtoMajor", c.Request().ProtoMajor).
-			Int("ProtoMinor", c.Request().ProtoMinor).
-			Msg("PROTOCOL-VIOLATION")
-		return echo.NewHTTPError(http.StatusBadRequest)
-	}
-
-	t := c.Request().Header.Get("x-sora-stats-exporter-type")
+func (s *Server) statsWebhook(c echo.Context) error {
+	t := c.Request().Header.Get("sora-stats-webhook-type")
 	switch t {
-	case "connection.user-agent":
+	case "connection.rtc":
 		stats := new(soraConnectionStats)
 		if err := c.Bind(stats); err != nil {
 			zlog.Warn().Err(err).Str("type", t).Send()
@@ -30,7 +21,7 @@ func (s *Server) collector(c echo.Context) error {
 			zlog.Warn().Err(err).Bool("simulcast", *stats.Simulcast).Str("type", t).Send()
 			return echo.NewHTTPError(http.StatusBadRequest, err.Error())
 		}
-		if err := s.collectorUserAgentStats(c, *stats); err != nil {
+		if err := s.rtcStats(c, *stats); err != nil {
 			zlog.Warn().Err(err).Str("type", t).Send()
 			return echo.NewHTTPError(http.StatusBadRequest, err.Error())
 		}